+++ test/View/Helper/NumberFormatTest.php
@@ -75,6 +75,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::DECIMAL,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1.234.567,891'
             ],
@@ -83,6 +84,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::DECIMAL,
                 NumberFormatter::TYPE_DOUBLE,
                 6,
+                [],
                 1234567.891234567890000,
                 '1.234.567,891235',
             ],
@@ -91,6 +93,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::PERCENT,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '123.456.789 %'
             ],
@@ -99,6 +102,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::PERCENT,
                 NumberFormatter::TYPE_DOUBLE,
                 1,
+                [],
                 1234567.891234567890000,
                 '123.456.789,1 %'
             ],
@@ -107,6 +111,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::SCIENTIFIC,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1,23456789123457E6'
             ],
@@ -115,6 +120,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::DECIMAL,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1 234 567,891'
             ],
@@ -123,6 +129,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::PERCENT,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '123 456 789 %'
             ],
@@ -131,6 +138,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::SCIENTIFIC,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1,23456789123457E6'
             ],
@@ -139,6 +147,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::DECIMAL,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1,234,567.891'
             ],
@@ -147,6 +156,7 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::PERCENT,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '123,456,789%'
             ],
@@ -155,8 +165,20 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
                 NumberFormatter::SCIENTIFIC,
                 NumberFormatter::TYPE_DOUBLE,
                 null,
+                [],
                 1234567.891234567890000,
                 '1.23456789123457E6'
+            ],
+            [
+                'en_US',
+                NumberFormatter::PERCENT,
+                NumberFormatter::TYPE_DOUBLE,
+                null,
+                [
+                    NumberFormatter::NEGATIVE_PREFIX => 'MINUS'
+                ],
+                -1234567.891234567890000,
+                'MINUS123,456,789%'
             ],
         ];
     }
@@ -164,27 +186,36 @@ class NumberFormatTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider currencyTestsDataProvider
      */
-    public function testBasic($locale, $formatStyle, $formatType, $decimals, $number, $expected)
+    public function testBasic($locale, $formatStyle, $formatType, $decimals, $textAttributes, $number, $expected)
     {
         $this->assertMbStringEquals($expected, $this->helper->__invoke(
             $number,
             $formatStyle,
             $formatType,
             $locale,
-            $decimals
+            $decimals,
+            $textAttributes
         ));
     }
 
     /**
      * @dataProvider currencyTestsDataProvider
      */
-    public function testSettersProvideDefaults($locale, $formatStyle, $formatType, $decimals, $number, $expected)
-    {
+    public function testSettersProvideDefaults(
+        $locale,
+        $formatStyle,
+        $formatType,
+        $decimals,
+        $textAttributes,
+        $number,
+        $expected
+    ) {
         $this->helper
              ->setLocale($locale)
              ->setFormatStyle($formatStyle)
              ->setDecimals($decimals)
-             ->setFormatType($formatType);
+             ->setFormatType($formatType)
+             ->setTextAttributes($textAttributes);
 
         $this->assertMbStringEquals($expected, $this->helper->__invoke($number));
     }